refactor(#451): model directory
